requirements.txt
pysha3 >= 1.0.2
